refactor(code-complexity): split score.rs into rules, rust, and score modules
- rules.rs: LanguageRules trait + associated types (NodeRole, NestingKind, ScoringUnit)
- rust.rs: Rust impl + its helpers (enclosing_impl_type, callee_name, scope_is_self)
- score.rs: domain types (Construct, LogicalOp, ContributorKind, etc.) + scoring algorithm + tests
