+++ crates/scute-core/src/code_complexity/mod.rs
@@ -25,6 +25,8 @@
 //! [`Evidence`](crate::Evidence) entries explaining what drives the score.
 
 mod check;
+mod rules;
+mod rust;
 mod score;
 
 pub use check::{CHECK_NAME, Definition, check};
